+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased](https://github.com/OpenLineage/OpenLineage/compare/0.7.1...HEAD)
 ### Added
 * Support for HiveTableRelation as input source in Spark integration [@collado-mike](https://github.com/collado-mike)
+* Add HTTP and Kafka Client to `openlineage-java` lib [@wslulciuc](https://github.com/wslulciuc), [@mobuchowski](https://github.com/mobuchowski)
 
 ### Fixed
 * GreatExpectations: Fixed bug when invoking GreatExpectations using v2 API [@collado-mike](https://github.com/collado-mike)